requirements.txt
requests==2.31.0
python-dotenv==1.0.0
discord.py==2.3.2
schedule==1.2.0
openai==1.3.0
pandas==2.1.4
numpy==1.24.3 